Tighten Binary/Unary Typechecking
Summary:
This does a few related things to ready the typechecker for changes to dynamic:

* Tightens up the type checker for binary and unary expressions, incorporating dynamic throughout. It is now tighter where it can be (e.g. shifts always yield ints), more restrictive where it should be (e.g. bitwise operations are only allowed on ints), standardizes certain behaviour better (e.g. the sum of any and int no longer depends on the order of their summation), and fixes missed cases (e.g. dynamic minus mixed is no longer num). It is also a little more table driven, so it should be easier to tweak in the future.

* Disallows array summation as this feature is losing support and doesn't play nicely with dynamic's typing rules.

* Removes tests that test no-longer-supported features (e.g. array addition and bitwise xor for booleans).

The basic rule for dynamic in expressions is that it will yield the most restrictive typing that the runtime could successfully give, excepting where that type is inconvenient for a user, in which case the type will be dynamic. So e.g. summing dynamic and int gives num, but dynamic + dynamic yields dynamic, since the runtime still supports summing arrays and num|array is an awkward type. This makes it so every successful usage of dynamic is acceptably typed and runtime type errors are contained to usages of dynamic; if the code type checks, you won't ever find that a non-dynamic result of using dynamic causes a type error at runtime when used later in the code.

Some particular changes to note:

* Using an any type in a binary or unary expression will always yield an any type. This will make it look like type coverage goes down, but will better reveal untyped data flow in a consistent manner. This change came about due to refactoring/bugfixing expression typing.

* Bitwise xor is no longer encouraged for logical xor. Instead, !== is.

* Some typing reasons are mildly different.

Future work: Give better reasons for typing.

Reviewed By: jamesjwu

Differential Revision: D8824094

fbshipit-source-id: fec2c8a77370a2a1f813c730f12bfd81159f5f40