Control number of threads using a runtime option
Summary:
The threaded implementation of debug-parser maintains
a separate state of libdwarf/libelf for each thread. This lib
maps the full binary into memory, which can be quite large and,
depending on the number of threads and hhvm binary size, will
exhaust all available memory and crash with an allocation error.

Reviewed By: markw65

Differential Revision: D9127053

fbshipit-source-id: c25034cf8dd4e11b04c322d511221982f010fb74