Add a build mode with wide tv_val<>
Summary:
This introduces a WIDE_TV_VAL build mode, where tv_val<> holds
separate pointers for the type and value. For now, all it does is cause a
performance regression, but it will allow us to start changing the internal
representation of arrays/objects/etc... to hopefully claw back the regression
and turn it into a win.

Reviewed By: markw65

Differential Revision: D8551301

fbshipit-source-id: a2fa616fffcb64c1b23bce04635a0b27f2479acc