HHVM Debugger: Flush outoging messages to client before shutting down
Summary: Currently the debugger discards any pending messages for the client before shutting down the debugger extension. But this is making it extremely difficult to unit test the debugger protocol logic reliably because the shutdown messages may or may not arrive depending on how quickly HHVM shuts down after running a test script.

Reviewed By: velocityboy

Differential Revision: D9125050

fbshipit-source-id: 883f0edf34fb8ffd53fd768b0a00367e2311ba9d