Fix deduplication of hover results.
Summary:
Right now we're deduplicating hover results after we flatten the various `hover_info` objects into their wire form as a list of snippets and markdown.  This means that in the case where we have

```
{snippet=A; addendum=B}; {snippet=A; addendum=B}
```

we'll flatten that to

```
A; B; A; B
```

and it won't get deduplicated!  This change moves deduplication so that it occurs before flattening, and we end up with

```
A; B
```

as we intended.

Reviewed By: arxanas

Differential Revision: D9142528

fbshipit-source-id: 0340b6e7a20c3276c81e452cd303bc38929ab491