Short Circuit Subtyping
Summary: Subtyping calls are expensive and becoming used more often for correctness purposes. Because the majority of subtyping is probably done on types that match exactly (as is almost definitely the case for all the instances I'm adding in with dynamic) this will cut short the expense. Now we will only go through the expensive simplification procedure if they aren't an exact match.

Reviewed By: jamesjwu

Differential Revision: D9152127

fbshipit-source-id: 522c7d0c3baf37a3a62dc8c9760dae7210091578