Forbid many expressions from being interpolated in strings
Summary:
The set of things you can interpolate in PHP strings is pretty arbitrary, and varies depending on which mode of interpolation you used.

A lot of the rules have been reversed engineered. The official PHP specification doesn't actually match up with reality. There's not an obvious set of rules dictating which things can be interpolated and when.

Reviewed By: pittsw

Differential Revision: D8986090

fbshipit-source-id: 63e3cb92c4348d710070c87fdf44b95916c274f4