call r-value qualified Future::then and remove unnecessary field.
Summary:
Overall plan to modify Future<T>::then to be r-value qualified and use Future<T>::thenTry or Future<T>::thenValue.

The goal is to disambiguate folly::Future and to improve type and lifetime safety of Future and its methods.

Move to r-value qualified call in hphp's watchman extension and remove class field that behaved as a temporary.

Post on the topic of improving safety of Future: https://fb.facebook.com/groups/fbcode/permalink/1726007767436055/

Reviewed By: yfeldblum

Differential Revision: D9177374

fbshipit-source-id: 7e8981f7cb5962bf74ed2b2d25f55da028569089