Make loosen_staticness() be recursive
Summary:
In HHBBC, when loosening staticness on a type, the staticness should be
loosened recursively. If we don't want to make staticness assumptions for a
type, we shouldn't be making staticness assumptions for any part of the
type. For example, when inferring a context sensitive function return type, we
want to loosen staticness from the returned value. This is because we may have
assumed constant propagation while analyzing the function inline, but we can't
actually do the constant propagation. However, if the function is async, the
returned type will be WaitH<Ty>. So, we want to loosen staticness on the inner
type, not just the outer.

This causes a (slight) regression in analysis because loosening staticness on a
SStr will drop any known constant value. This will be addressed in a follow-up
diff.

Reviewed By: mxw

Differential Revision: D9177816

fbshipit-source-id: 2f35ff29dbebede5331abe8a8ed5ad85463c1824