HHVM Debugger: Print exception info when break on all exceptions is enabled
Summary:
Previously we defaulted to a mode where it warned in the console on notice/exception but did not break. Now we only offer an exception break filter for break all or break none (because the VM doesn't let us detect if the exception is going to be handled or not). It's really annoying not to be able to see the exception message when the debugger stops at the exception breakpoint when this mode is enabled - you can't see the message until you step/resume and let the VM write to stderr.

This diff prints the message to the debugger console right away.

Reviewed By: velocityboy

Differential Revision: D9162124

fbshipit-source-id: f15dc1368c79ca4463d8204bcb64d43800ed44c7