Plumb through Watchman changed merge base info to Restart_server saved state target
Summary:
Story time.

In an Informant-directed restart, Watchman sends a notification to the
Informnat specifying "n" as the new mergebase. The Informant finds a Saved State,
provides the Saved State target to the new Server. New Server starts a
Watchman subscription, then loads that Saved State, then queries
hg for files changes since that SS, then asks hg what the current mergebase
is - we should expect this query to also report "n" as the mergebase.

Sometimes during rebases it does not.

It is rare, but it gives a bad experience when
it happens. I am not able to repro it locally, but I have observed it
in the wild while combing through server logs and event tables and hg reflogs.

This diff threads through that "n" from the Informant to the Server.

Also, when "n" is misreported, the hg query for
"files changed since that SS" also misreports. This query also implicitly
relies on the current node in hg to be working correctly, so we want to
avoid this too.

A followup diff in this stack will change the strategy for calculating
"dirty files" to avoid that query.

Currently, dirty files is defined by the singly query
"Hg.files_changed_since_rev saved_state_rev" which returns files
changed between that revision and the current repo state (which gets messed
up).

This will be changed instead to "files changed between Saved State's
revision and the new mergebase". It would seem we still need to query
mercurial with "hg.files_changed_since_rev mergebase_rev", causing
us to implicitly rely on hg's current state to work properly - but
actually this information is provided to the Informant inside the
Changed_merge_base notification. So we just need to thread that through!

Finally, in order for us to capture any other files that change
during this entire journey to init, we need to thread through the
Watchman clock, so that the new subscription that's started in the new
Server can begin at that clock.

Please see ServerMonitorUtils.watchman_mergebase for new
data that is threaded through to the new Server. Currently it is completely
unused in the Server. That will be done in a followup.

tl;dr Sometimes hg goes weird, causing two queries (that Hack does during init) to
report false results: 1) what is current mergebase 2) files changed since Saved State's revision. This diff threads through more data from the Informant to the newly spawned Hack server to avoid both of these queries. i.e., use Watchman subscription notifications all the way through instead of a weird mix of Watchman notifs and hg queries which apparently have conflicting ideas of the state of the world.

Reviewed By: dabek

Differential Revision: D8431351

fbshipit-source-id: 5f610f18de7d20114178f4ffaa22c91ad240626b