Use watchman mergebase during Server init
Summary:
See D8431351 for the full story. It is required reading.

This is the big payoff.

During initialization, we use the watchman mergebase clock provided by
the Monitor to start the watchman subscription - this provides a stream
of file changes that continue off from when the Changed_merge_base
notification was given to the Monitor.

We also change the computation of dirty_files to be the files changed
since the Saved State's SVN Rev and the mergebase we moved to (mergebase
as specified by the Monitor). That dirty_files list will be added to
"files changed during init" (as it always has in the past) inside
ServerInit; since the watchman subscription is started off with the
clock given in the Changed_merge_base notification, this will effectively
capture all the files that changed since that Changed_merge_base notification
all the way up to ServerInit (including all the time it took to find
and prefetch a saved state, kill the previous server, and launch a new one).

Reviewed By: dabek

Differential Revision: D8554935

fbshipit-source-id: 58cebdee92f0b85cda112e42eac03f026ba36606