Omit sign-extensions for narrow comparisons
Summary:
We currently unconditionally insert a sign-extension when we compare
bytes or 16-bit words; however, for equality checks, this isn't strictly
necessary. This runs annotateSFUses to flag the instruction that generates the
StatusFlags with the bits that are actually checked. If this value is just Z
that means we are using condition code EQ or NE and can leave the sign
extension out.

Vixl also tries to be clever by emitting a `cmn` instruciton and negating the
immediate if we try to `cmp` with a negative immediate; however, since we use
early-truncated Reg{8,16}s, I need to force vixl to not negate the immediate
for these cases; while this breaks other use-cases (any signed comparisons that
try to use `cmpb` and friends), they will all be lowered differently by
`lower(cmp{bw}[i])` and won't use the "weird" flavors of these unless the
comparison is narrow.

Reviewed By: ottoni

Differential Revision: D9057163

fbshipit-source-id: a05a63fb12e2c71e622c37ec46b0af1225b378ed