Add 'exact' field to TypeStructures representing the 'this' type
Summary:
The `this` type is special, in that it is an exact type: when testing against it, the value must be an instance of exactly the late static bound class (not a subtype).

For `is`-expressions, by the time we check against a type structure, `this` types are resolved into their actual classes, so we don't know if we must check that the value is a subtype or the exact type.

This adds an "exact" field to the type structure of classes/interfaces to indicate whether it was resolved from a `this` typehint.

Reviewed By: mofarrell

Differential Revision: D9109902

fbshipit-source-id: 8c4fc40eb7db80dd39acadecc50215c82b2baea7