Log function types with Tany as separate instances
Summary:
We test equivalence of function types for the purposes of logging legacy lambdas using `is_sub_type`. Unfortunately this conflates uses of `Tany` with any type, leading to unsound suggestions (e.g. it might lead to a user annotating a lambda with a type that isn't satisfied at runtime).

To mitigate this, split function types into those containing `Tany` and those that don't.

Reviewed By: manzyuk

Differential Revision: D9194915

fbshipit-source-id: 6259b20af61130bf25fef995e42d49835d7f21f6