Refactor Decl_init_check to prepare for __LateInit
Summary:
Currently, when checking that all of a class's properties are initialized, we only have access to the property names. For `__LateInit`, we'll want access to the actual property element, which will contain a flag to tell us if it's late init.

This genericizes the functions in `Decl_init_check` to take `add_prop` functions instead of assuming that the return value is a Set. In a followup diff, this will be changed so that `Nast_init_check` receives a Map.

Reviewed By: jamesjwu

Differential Revision: D9130827

fbshipit-source-id: d126b7b0067fba0c005eae309316d8f98660feb3