Add prop type-hint enforcement for remaining cases
Summary:
Add logic to enforce property type-hints for the remaining cases. This includes
the various places we can set a value in a property, as well as deserializers
which can create objects. Properties with type-hints cannot be unset nor bound
to a ref, so forbid those operations as well.

The main complexity is with the IncDec and SetOp operations, since those typical
operate on the value in place, but they can change the type. In those cases, we
must perform the set-op on a temporary copy, verify its type, then move it back
into the property. In a few cases we can infer the type won't change and its
safe to do the operation in place.

BindS and VGetS would require new HHIR ops to implement with the JIT. However
these are very rare operations (and refs are slowly being removed), so instead
make them interp-one ops and have the interpreter deal with it.

With type-hint enforcement fully enabled, the JIT will now take advantage of the
type-hints to provide a lower bound on the property's type.

Plumb tv_rval thru type-constraint more.

With this diff, property type-hint enforcement is now fully implemented.

Reviewed By: swtaarrs

Differential Revision: D8832054

fbshipit-source-id: 332446a330715d4f1662310c6a3b059df020cd34