Implement <<__LateInit>> properties
Summary:
Add a new attribute for properties called <<__LateInit>>. This attribute marks
that a property does not require an initial value (nor being set in the
constructor) and starts out in an uninitialized state. If the property is read
in this uninitialized state, an exception is thrown. Once set, the property
behaves normally otherwise. A property can be moved back to its uninitialized
state by using unset (though only if there's no type-hint). A property can be
tested if its uninitialized by using isset() or empty() (this does not attempt
to access the value, thus it does not throw an exception). The use case for this
feature is if you have a property (static or otherwise) which is awkward to
initialize sanely initially, but you don't want to use a placeholder value and
you want runtime verification that you're not using the property before its
ready.

The implementation is straightforward. The property starts out as Uninit, though
it does not get set in 86pinit or 86sinit. When loading the property in a read
context, we'll check if the property is Uninit, and if it is, throw. This
required a few fixups in HHBBC since a <<__LateInit>> property which is never
set will have a Bottom type, and HHBBC previously asserted that properties were
never Bottom.

Reviewed By: markw65

Differential Revision: D9072061

fbshipit-source-id: 8b92368704f2892d3f8d4775657f76ef7e0c3302