FuncTable part 2: assembler & verifier
Summary:
Use UnitEmitter.m_nativeFuncs in verifier. (one simple callsite).

Plumb nativeFuncs through assembler apis to the points where
the assembler creates UnitEmitters.

Relax the logic for setting native/builtin flags in the assembler;
allow native decls outside systemlib, however if systemlib is
already initialized, fail with an error if a native func mapping is
not found in the provided table.  This changes the failure mode
for userspace code trying to define a native function, but it still
fails.  Now, it's possible to define native functions after systemlib
has initialized, but such functions will not be given AttrBuiltin,
AttrSkipFrame, or AttrMayUseVV.

Reviewed By: markw65

Differential Revision: D9083506

fbshipit-source-id: f6c616fe34982265d1e7c58e0872deeb2953a55b