Deal with SCOPE_ASSERT_DETAIL in the crash-reporter
Summary:
Now that the crash reporter is setup to recover from seg faults while
it runs, run the AssertDetails one at a time, so that we get the
output from any that don't crash. This also means we get to see the
details for any signal (and in particular it no longer depends on
assertx vs assert).

Reviewed By: binliu19, alexeyt

Differential Revision: D9174159

fbshipit-source-id: 590df60edc05b0309850dc23ce43917359ff247d