Add Unsafe_block to NAST and TAST
Summary:
The `// UNSAFE` comment, which makes any subsequent statements in the containing block invisible to the typechecker, is represented with an `Unsafe` statement in the AST. In naming, we currently delete `Unsafe` statements and any statements following them in a block, and don't represent the contents of unsafe blocks in any way in the named and typed ASTs.

This would be reasonable if the typechecker were the only consumer of named and typed ASTs, but other static analysis tools would like the ability to analyze unsafe blocks. This change introduces an `Unsafe_block` variant to the `Aast.stmt` type, and wraps any statements following an `// UNSAFE` comment in an `Unsafe_block`.

Reviewed By: fahndrich

Differential Revision: D9155930

fbshipit-source-id: 976b298402a1ccd55d673d7a9d22d992bb0df392