Remove ClassIdAnnotation
Summary: ClassIdAnnotation was added in {D6617749} to track the resolved types of `class_id`s in the TAST. At the time, a `class_id` did not have an associated position, so it would not have been easy to reuse the `ExprAnnotation.t` type. Now that {D8335742} has landed, `ExprAnnotation.t` and `ClassIdAnnotation.t` are always the same type, and we don't really need to distinguish them anymore.

Reviewed By: pittsw

Differential Revision: D9198722

fbshipit-source-id: 1a5578f06b3d66953eb6365027f6f9dc291d1025