Enforce that the enum constraint, if given, is a subtype of arraykey
Summary:
Currently it is possible to specify any `as` bound on an enum (that is a supertype of the underlying enum type).  For example:
```
enum N: int as ?int {
  Z = 0;
}
```
is accepted.  However, in the type-checker we unconditionally treat enums as subtypes of arraykey, which leads to surprising effects:
```
function test1(): arraykey {
  return N::Z;
}
```
passes but
```
function expect_int(int $_): void {}

function test2(): void {
  expect_int(N::Z);
}
```
does not.  Furthermore, changing `arraykey` to `int` in `test1` makes it fail the type-checker too.

We should really enforce that the `as` bound in the definition of an enum is always a subtype of `arraykey`.

Reviewed By: kmeht

Differential Revision: D9220264

fbshipit-source-id: df45b5e96606cc006a1218a4760db40d6af5e2ad