Dynamic in expressions
Summary:
This provides typing rules for binary and unary expressions that include dynamic. It also removes the typing rules for array addition, as that feature is losing support and it doesn't behave well with dynamic.

This does not currently provide typing rules for the binary operation of null coalesce, neither the unary operations of ref, silence, nor increments/decrements. (Note, Hack does not have a logical xor operator aside from `xor`, which is ignored due to evaluation precedence issues.)

Reviewed By: andrewjkennedy

Differential Revision: D8737392

fbshipit-source-id: dbaac3e6e03026ba3af7b84f44247998b0703576