Implementing Dynamic Coercion
Summary:
This diff does a few related things for most of the functionality of changing `dynamic` to use our new notion of type coercion:

* Makes it so the `dynamic` type is no longer a top-like type.

* Allows types to be coerced to `dynamic` for property assignments, property initializations, returns, array assignments, function arguments, xml, and most kinds of array-like keys (some strictly require literals instead).

We also needed to consider cases where we get super and subtypes of `dynamic`, either because such was given as a generic constraint, or because of union or intersection types (although the latter doesn't exist yet). The rule here is that any type coerces to any supertype of `dynamic`, but values of such supertypes will lose the flexible usage that `dynamic` had. On the other hand, values of subtypes of `dynamic` can be used exactly as dynamic could be (in line with subsumption) but cannot generally be coerced to.

Some other changes related to `dynamic`:

* Since `dynamic` is no longer top-like, extension cannot use it as such. An interface's function returning `dynamic` cannot be implemented with a function that returns `int`.

* Performing a `yield from` on a `dynamic` in a function body requires that function's return type to be `dynamic`.

* Made `awaitable<T>` coerce to `awaitable<dynamic>`

Future work:

* Add logging for `dynamic` coercions.

* Give better reasons for typing.

* Class extension from `dynamic`.

* Let expressions were left untouched, so they might want to use coercion too.

* Use coercive constraints in generic resolution in addition to subtype.

Reviewed By: jamesjwu

Differential Revision: D8787271

fbshipit-source-id: 5568362e49e0f5d3c38d8745da566fb74cc3ea46