Better Reasons For Hack Binary/Unary Typing
Summary:
When refactoring the typechecker for binary/unary expressions, it was noticed that the reasons given for typing were a little scattered in where they pointed and the kind of info they gave. This upgrades them.

In particular:

* arguments typed dynamic will explicitly say why instead of e.g. "was used in an arithmetic expression"

* integer arithmetic is distinguished from more general arithmetic

* floats and nums name the responsible argument and give the root cause of that argument's type (previously it just returned as if the reason was exactly that of the root cause)

Also the array addition reason was removed because array addition is no more.

Reviewed By: jamesjwu

Differential Revision: D9253825

fbshipit-source-id: 5022c74c300244db26d5783267a2d8ebcd78cf75