HHVM Debugger: Fix error spew when evaluating in hover or watch context
Summary:
When the debugger client evaluates in the "hover" or "watch" contexts, if the expression cannot be evaluated, we should not be throwing PHP errors and printing a bunch of spew.

This diff fixes a few issues:
1. Do not print an exception to the console for eval hover errors
2. In Nuclide, don't show a data tip with "N/A" or an expression evaluation error when the user hovers over something that is undefined, a keyword, or not valid code.
3. The SilentEvaluationContext shouldn't be setting the ExecutionContext output buffer to nullptr - this is causing the error not to be suppressed, but instead written to the TLS request transport fd. If the debugger is in webserver attach mode, this causes the error to be written to the browser, which corrupts the webpage output!

Note re #3: As best I can tell, hphpd has the same bug due to null'ing the transport buffer during evalaution, but since the CLI doesn't have hover events, it's not noticable.

Reviewed By: velocityboy

Differential Revision: D9171487

fbshipit-source-id: c83ddde333cb43f40a9c56dd3bde0cf94dfdde3d