Fix/add tests for comparing classes with special toArray behavior.
Summary:
ArrayObject, ArrayIterator, and SimpleXMLElement have special behavior in ObjectData::toArray, but when comparing them we weren't dealing with this (except, bizzarely, when testing equality on ArrayObject), and we didn't have tests comparing 2 instances of these classes to each other.
Our behavior for all 3 was different from PHP, and the behavior for ArrayObject and ArrayIterator made no sense. However, our behavior for SimpleXMLElement is more useful than the PHP behavior (which essentially treats them as non-comparable, similar to closures). https://3v4l.org/vlQoe
Fix the behavior for ArrayObject and ArrayIterator to match PHP, keep our behavior for SimpleXMLElement, and add test coverage for this.

Reviewed By: ricklavoie

Differential Revision: D9262552

fbshipit-source-id: 865cb350c4da2ef5b6571d5b51d456f84a1ee549