Stop using ObjectData::toArray in ObjectData::equal.
Summary:
We don't need to iterate properties in any particular order* to determine if 2 objects are equal, so we can iterate the properties in memory order (and use array equality on dynamic prop arrays if they're there).
Be careful when handling objects with special toArray behavior (tests added on the previous diff) and to avoid running out of stack on recursive objects (tests added in this diff).

Reviewed By: ricklavoie

Differential Revision: D9262553

fbshipit-source-id: f69766b54ba64a6f0c277b43ae88109e81852bc6