Stop doing nonsensical refcount operations in objpropf
Summary:
Objprof casts objects to arrays in order to iterate their properties, so it adjusts the resulting refcounts down by 1 when it does so.
It was also incorrectly adjusting refcounts down when iterating static properties (where no incref had happened that needed to be compensated for). Fix this.
Finally, it was checking that the refcount after adjustment was > 0 before divinding by it. Aside from the above bug with static properties (and one bit refcount mode, which didn't exist when the code was written) this makes no sense. Fix this: in one bit refcount mode, since we don't know the number of places referring to a value, always count the entire size of the value without dividing. In normal mode, assert that the adjusted refcount was > 0 instead of checking for it.
I also updated the list of failing slow tests in one bit refcount mode; several previously failing tests are now passing (unrelated to this diff).

Reviewed By: markw65

Differential Revision: D9310988

fbshipit-source-id: 75fb49d258919cfd82a6e8cd26799b3d0e9e5bf3