Stop using ObjectData::toArray in objprof
Summary:
Create a helper for iterating object properties (in memory order) without copying them all into an array first.
Use it in objprof, and stop subtracting 1 from refcounts to compensate for it.
Fix a bug in objprof's getObjSize where we would forget to pop the object off the val_stack when we returned early in the collections case.
Remove a bogus comment from hphp/runtime/vm/class.h: Class::Prop::name is always the name of the PreClass Prop, it's not set to "" when the Prop is inaccessible.
Deleted test/slow/objprof/objprof-refs.php. It was testing that we didn't crash by using strings after free, which could happen if an object on which we called toArray had special toArray behavior, and returned an array containing temp strings rather than ones pointed to from the object we were inspecting. That's not possible any more so the test is no longer relevant.

Reviewed By: markw65

Differential Revision: D9310989

fbshipit-source-id: 07cf015bc4ee6604ba25647613bfcd0ffa7fc395