Adding signature algorithm to output of openssl_x509_parse
Summary:
openssl_x509_parse doesn't include the certificate's signature algorithm, which we need for some functionality in www. It's already in the cert info, so we just need to decode add it to the map of parameters.

Right now there's two possible implementations in here. I'm not sure which is the most kosher, so I'd appreciate feedback on that.

Reviewed By: fredemmott

Differential Revision: D8942393

fbshipit-source-id: 56ba637bf607bec75b227830a90c117e5d13b423