Deprecate legacy Aast visitors
Summary: Now that we generate NAST/TAST visitors using the visitors PPX extension, we should start migrating away from the hand-written visitors. This change marks their base class as deprecated and moves its definition out of the AnnotatedAST functor and into the Nast module, since it is currently only used on the Nast.

Reviewed By: kmeht

Differential Revision: D6481756

fbshipit-source-id: 9bf592cac1b0df548078ea0a6a9bd9e4a1cb30f2