Refactor hackfmt main to have only one Logger.exit call
Summary:
Some refactoring I did while investigating the problem that led to D9011103. When I originally added this logging, I wasn't sure how to do this nicely because I was not so conscious of the fact that `try` is an expression.

Note that this causes us to log 0 for the exit code of a successful exit rather than null. This seems like an improvement to me.

Reviewed By: arxanas

Differential Revision: D9351400

fbshipit-source-id: 5a990820bd5b5dfa8cde9b2b38806341638cfccf