Refactoring verify.py so we could verify idempotence
Summary: Changing the type from Failure to Result and letting the run_test_program function return all the results instead of just the failures. This is in order to be able to add the idempotence test step that can run on top of the results obtained from running the test program.

Differential Revision: D9260112

fbshipit-source-id: 64eb7f03b9dc168b74a51f9973400bcae18063fa