Rename Variant::asCell and String::asCell to toCell
Summary:
The naming convention on Variant is that toFoo assumes the variant
holds a Foo, or a reference to Foo, and unboxes if necessary, while
asFoo assumes it contains an unboxed Foo - so asCell was misnamed.

On String, there's no unboxing issue, but all the other conversions
are named toFoo, so rename asCell for consistency.

Reviewed By: swtaarrs

Differential Revision: D9350750

fbshipit-source-id: 04c847d4a8fdd382a62e493f6c20bf3e27589bc5