Get rid of architecture-specific adjustMetaDataForRelocation
Summary:
Merge the architecture-specific versions into the main function.  They can all
do the same thing and, if a backend doesn't use any of the vectors/maps, it'll
just skip over.  Indeed, the architecture-specific versions were almost doing
the same thing already.  This diff is mostly code motion, but it fixes a
potential bug on PPC: it uses smashableLocations, but the code updating that was
only in the ARM-specific function.

Reviewed By: jjgriego

Differential Revision: D9356582

fbshipit-source-id: 57c208705bbbd4ae4f4d33f98124f9f2517ebcdb