Catch infinite recursion in genLayout in gen-type-scanners
Summary:
Add logic to detect infinite recursion when generating layouts in
gen-type-scanners. In most cases, this is either a bug in gen-type-scanners, or
an indication of invalid debug information. Once case where it can happen
legitimately, however, is when testing if a custom type-scanner should be
called. If a custom type-scanner has a type parameter, we'll attempt to generate
the layout for that type parameter, and see if it has any non-trivial
actions. One example where this can result in infinite recursion is if we have a
req::vector containing a type which ultimately contains the same req::vector. By
throwing a LayoutError when infinite recursion is detected, this is dealt with
automatically, since we already suppress LayoutErrors in this case.

Reviewed By: edwinsmith

Differential Revision: D9364823

fbshipit-source-id: 5cde86be9d5fc101db420ae21ead80d1ac3025ff