Forcibly inline all tv_val<>-related functions
Summary:
I recently discovered that some of the calls to HPHP::as_tv() from
this file aren't being inlined, which could easily explain the regression I'm
seeing in some of my pending diffs. Since tv_val is meant to be a zero-cost
abstraction for a real pointer, inline and flatten all functions dealing with
it (as_tv(tv_lval) specifically is analogous to dereferencing a TypedValue*).

Reviewed By: mxw

Differential Revision: D9187815

fbshipit-source-id: 08d3fb213aa1ca187ad5bc8816e121b4ea8106be