Expand diff formatting intervals outside formatting range when range starts at atom boundary
Summary:
We currently fail to add leading indentation when formatting a diff which adds an unindented line:

```
 $x = builder()
+->setFoo(3)
   ->setBar(3)
   ->build();
```

In diff mode, we take the line range of the newly added lines, expand the ranges to atom boundaries so as to include any surrounding whitespace, then format those ranges. However, in this case, that isn't sufficient--since the range of the newly added lines begins at the start of the `->` token, it is already at an atom boundary, and we choose not to expand it further.

Since --diff mode expands ranges to atom boundaries and does not print whitespace outside those boundaries, we do not add any indentation before the `->` token. This isn't the intended behavior--we should expand to the nearest atom boundaries //not contained by the range//, expanding the start of the formatting range to include the `)` token on the first line, so that we include the whitespace preceding the `->` token.

Reviewed By: arxanas

Differential Revision: D9210697

fbshipit-source-id: 4e89b9ee8538d4210c077983d5cdb91dd6660abd