Make verify.py and review.sh sensitive to trailing newlines
Summary:
Currently, neither of these scripts will detect when the expected output has a trailing newline, but the actual output doesn't (or vice versa). In the case of verify.py, it is an explicit decision called out in a comment (although I don't know the rationale). In the case of review.sh, it is a consequence of the fact that `diff` seems to be aliased to `'diff -wbBdu'` on devservers by default.

In order to add a test case verifying that hackfmt --diff adds a trailing newline to files, I would like these scripts to detect when a trailing newline is present in one file and not the other. This change stops adding a trailing newline to the actual output in verify.py, and uses `command` to bypass the `diff` alias in review.sh.

Reviewed By: 2BitSalute, arxanas

Differential Revision: D9370078

fbshipit-source-id: 4c721d605ddab8570527b0c376d296d883b3e554