Ensure trailing newline in --diff mode when last line was modified
Summary:
When the last line in a file has been modified (most commonly, when a new file has been added), --diff mode currently fails to print a newline at the end of the file. This is because --diff mode takes the ranges of the modified lines, expands them to the nearest atoms outside the range, and then formats those ranges with no surrounding whitespace. Formatting the last token in the file with no surrounding whitespace means we fail to print the trailing newline!

Perhaps there's a more principled way of doing this--maybe we could have some kind of special atom representing the end of the file, and then the trailing newline would be included in the range after expanding to atom boundaries. Or maybe the line splitter could detect when the range includes the end of the file, and print trailing whitespace in that case. But neither of those seem much better to me than the simple solution, which is to just print a newline character at the end of the file if there isn't one there already.

Reviewed By: arxanas

Differential Revision: D9370114

fbshipit-source-id: aad108495b26ecb5f4754c46edffe5db4f31fdcb